Stop replacing the project name in package.json (#40)
We've moved away from using the ${PROJECT} and ${DESCRIPTION}
replacement strings in Pulumi.yaml and package.json, primarily so that
`pulumi new <url>` and `pulumi up <url>` can be used with examples that
don't use these replacement strings (allowing such examples to be
usable/buildable on their own without any templating).

Instead of doing a search/replace of these replacement strings, v0.15.2
of the CLI now overwrites any existing `name` and `description` values
in Pulumi.yaml and saves the file, and doesn't do anything to update the
`name` specified in package.json. The replacement strings can be used,
but they are no longer required with this version of the CLI.

However, v0.15.0 of the CLI only does the search/replace behavior, so we
must keep ${PROJECT} and ${DESCRIPTION} in Pulumi.yaml in the templates
in pulumi/templates as long as we want v0.15.0 of the CLI to continue
working with our templates. (Earlier versions of the CLI before v0.15.0
download templates from S3, and therefore aren't affected by any changes
we make to templates in this repo unless we force updates to the
templates in S3).

But there's no need to keep ${PROJECT} in package.json. Since we decided
we're no longer going to make an effort to update the project name in
package.json, we can simply remove the replacement string there.

This works around an issue where we currently allow project names that
can have characters (such as `/`) that NPM doesn't allow for the `name`
in package.json.