Fix Python template tests
The integration test package by default uses Pipenv to manage virtual environments for Python test projects. As of Pulumi 2.4.0, `pulumi new` will automatically create and manage a virtual environment itself, which is causing these template tests to fail because Pulumi is trying to manage a virtual environment created in the context of a Pipenv virtual environment. To address, we need to set the new `UseAutomaticVirtualEnv` test option, which prevents the use of Pipenv.