Update aws-go dependencies
It looks like the PRs that automatically update Go dependencies haven't been merged in a while.

To make it easier to get started trying the Go generics preview, update the dependencies for the aws-go template to the latest versions, along with the Go 1.20 (Go 1.18 and 1.19 are no longer supported). That way, users trying out generics, only have to update the `pulumi-aws` dependency to the preview branch, and don't have to update the core Pulumi SDK to a more recent version.