always use the latest pulumi dev version in CI
Similar to https://github.com/pulumi/examples/pull/1535, we would like
to use the latest pulumi dev version here instead of just the latest
version.

In face in some places there was some (very outdated and
non-functional) code to attempt that.  Instead of the latest dev
version it downloaded 3.40.0-alpha.1663101047 as of the time of this
writing, which is less than ideal.  Always use the dev version here
unless we explicity request the "latest" version.