Skip ovh-java test
There's an issue with this template. We're likely going to reduce the number of templates we support out of the box to only those that are most widely used, and likely ovh will be moved out (can be hosted elsewhere). Therefore, it's not worth spending a lot of time fixing up this template, so skip testing it for now.